development/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_slug}}"
version = "1.0.0"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.author_name}}"]
license = "{{cookiecutter.open_source_license}}"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
bcrypt = "^3.2.0"
Flask = "^3.0.2"
flask-cors = "^4.0.0"
Flask-SQLAlchemy = "^3.1.1"
flask-jwt-extended = "^4.6.0"
flask-caching = "^2.1.0"
flask-smorest = "^0.43.0"
flask-bcrypt = "^1.0.1"
flask-limiter = "^3.5.1"
flask-testing = "^0.8.1"
Flask-WTF = "^1.2.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
pytest = "^8.0.0"
coverage = "^7.4.1"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
httpx = "^0.26.0"
pendulum = "^3.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
